@@ -4,7 +4,7 @@ Blockly.Msg.ADD_COMMENT = "Magdagdag ng komento";
 Blockly.Msg.ADD_PARAMETER = "Magdagdag ng parameter";
 Blockly.Msg.CANCEL = "Kanselahin";
 Blockly.Msg.CHANGE_VALUE_TITLE = "Baguhin ang value:";
-Blockly.Msg.CLOSE = "Close";
+Blockly.Msg.CLOSE = "Isara";
 Blockly.Msg.COLLAPSE_ALL = "I-collapse ang mga blocks";
 Blockly.Msg.COLLAPSE_BLOCK = "i-collapse ang block";
 Blockly.Msg.COLOUR_BLEND_COLOUR1 = "kulay isa";
@@ -26,7 +26,7 @@ Blockly.Msg.COLOUR_RGB_TITLE = "kulayan ng";
 Blockly.Msg.COLOUR_RGB_TOOLTIP = "Gumawa ng kulay ayon sa tinugon na dami ng pula, berde, at asul. Lahat ng halaga ay kinakailangang nasa pagitan ng 0 hanggang 100.";
 Blockly.Msg.CONFIRM_CREATE_VARIABLE = "I-create";
 Blockly.Msg.CONFIRM_DELETE_FUNCTION_MESSAGE = "Babala! Sigurado ka ba na gusto mo i-delete ang %1?";
-Blockly.Msg.CONFIRM_RENAME_VARIABLE = "Rename";
+Blockly.Msg.CONFIRM_RENAME_VARIABLE = "Baguhin ang Pangalan";
 Blockly.Msg.CONTRACT_EDITOR_HEADER = "Kasangkapan ng Pag-design";
 Blockly.Msg.CONTROLS_FLOW_STATEMENTS_HELPURL = "https://code.google.com/p/blockly/wiki/Loops#Loop_Termination_Blocks";
 Blockly.Msg.CONTROLS_FLOW_STATEMENTS_OPERATOR_BREAK = "putulin n paulit-ulit";
