@@ -46,7 +46,7 @@ Blockly.Msg.CONTROLS_FOR_TOOLTIP = "U ovom bloku varijabla %1 poprima vrijednost
 Blockly.Msg.CONTROLS_IF_ELSEIF_TOOLTIP = "Dodaje uvjet za Ako blok.";
 Blockly.Msg.CONTROLS_IF_ELSE_TOOLTIP = "Dodaj posljednji uvjet, za sve preostale slučajeve, u blok \"ako\".";
 Blockly.Msg.CONTROLS_IF_HELPURL = "http://code.google.com/p/blockly/wiki/If_Then";
-Blockly.Msg.CONTROLS_IF_IF_TOOLTIP = "Dodaj, ukloni ili razmjesti dijelove da preoblikuješ ovaj \"ako\" blok.";
+Blockly.Msg.CONTROLS_IF_IF_TOOLTIP = "Dodaj, ukloni ili razmjesti dijelove da preoblikuješ ovaj Ako blok.";
 Blockly.Msg.CONTROLS_IF_MSG_ELSE = "inače";
 Blockly.Msg.CONTROLS_IF_MSG_ELSEIF = "inače ako";
 Blockly.Msg.CONTROLS_IF_MSG_IF = "ako";
@@ -63,7 +63,7 @@ Blockly.Msg.CONTROLS_REPEAT_TOOLTIP = "Ponavlja određene naredbe nekoliko puta.
 Blockly.Msg.CONTROLS_WHILEUNTIL_HELPURL = "http://code.google.com/p/blockly/wiki/Repeat";
 Blockly.Msg.CONTROLS_WHILEUNTIL_OPERATOR_UNTIL = "ponavljaj dok ne bude";
 Blockly.Msg.CONTROLS_WHILEUNTIL_OPERATOR_WHILE = "ponavljaj dok je";
-Blockly.Msg.CONTROLS_WHILEUNTIL_TOOLTIP_UNTIL = "Dok uvjet nije ispunjen, onda izvrši neke naredbe.";
+Blockly.Msg.CONTROLS_WHILEUNTIL_TOOLTIP_UNTIL = "Dok uvjet nije ispunjen, izvršava zadane naredbe.";
 Blockly.Msg.CONTROLS_WHILEUNTIL_TOOLTIP_WHILE = "Dok je uvjet ispunjen, izvršava zadane naredbe.";
 Blockly.Msg.DEFINE_HEADER_DEFINITION = "Definicija";
 Blockly.Msg.DEFINE_FUNCTION_DEFINE = "Definirati";
@@ -78,13 +78,13 @@ Blockly.Msg.DUPLICATE_BLOCK = "Dupliciraj";
 Blockly.Msg.ENABLE_BLOCK = "Omogući blok";
 Blockly.Msg.EXAMPLE = "Primjer";
 Blockly.Msg.EXAMPLE_DESCRIPTION = "Definira primjer s očekivanim i stvarnim ponašanjem";
-Blockly.Msg.EXPAND_ALL = "Rastegni blok";
-Blockly.Msg.EXPAND_BLOCK = "Rastegni blok";
+Blockly.Msg.EXPAND_ALL = "Proširi blok";
+Blockly.Msg.EXPAND_BLOCK = "Proširi blok";
 Blockly.Msg.EXPECTED = "rezultat";
 Blockly.Msg.EXTERNAL_INPUTS = "Vanjski unosi";
-Blockly.Msg.FUNCTION_CREATE = "Stvorite funkciju";
+Blockly.Msg.FUNCTION_CREATE = "Napravi funkciju";
 Blockly.Msg.FUNCTION_EDIT = "uredi";
-Blockly.Msg.FUNCTION_NAME_LABEL = "Imenujte svoju funkciju:";
+Blockly.Msg.FUNCTION_NAME_LABEL = "Imenuj svoju funkciju:";
 Blockly.Msg.FUNCTION_DESCRIPTION_LABEL = "Što bi vaša funkcija trebala učiniti?";
 Blockly.Msg.FUNCTION_PARAMETERS_LABEL = "Koje parametre ima tvoja funkcija?";
 Blockly.Msg.FUNCTION_HEADER = "Funkcija";
