@@ -1863,7 +1863,7 @@ StudioApp.prototype.configureDom = function (config) {
     // Temporarily attach an event listener to log clicks
     // Logs the type of app and the ids of the puzzle
     var videoThumbnail = document.getElementsByClassName('video_thumbnail');
-    if (videoThumbnail[0]){
+    if (videoThumbnail[0] && (config.app === 'gamelab' || config.app === 'applab')){
       videoThumbnail[0].addEventListener('click', () => {
         firehoseClient.putRecord(
           'analysis-events',
