@@ -1,6 +1,15 @@
 import React from 'react';
-import ManageStudentsTable from './ManageStudentsTable';
+import {UnconnectedManageStudentsTable} from './ManageStudentsTable';
 import {SectionLoginType} from '@cdo/apps/util/sharedConstants';
+import {combineReducers, createStore} from 'redux';
+import manageStudents from './manageStudentsRedux';
+import {Provider} from 'react-redux';
+
+const initialState = {
+  loginType: '',
+  studentData: {},
+  sectionId: null,
+};
 
 // Student names out of alphabetical order to demonstrate
 // sorting functionality in the storybook
@@ -17,6 +26,7 @@ const passwordAccountData = [
     secretPictureName: 'wizard',
     secretPicturePath: '/wizard.jpg',
     sectionId: 53,
+    isEditing: false,
   },
   {
     id: 2,
@@ -30,6 +40,7 @@ const passwordAccountData = [
     secretPictureName: 'wizard',
     secretPicturePath: '/wizard.jpg',
     sectionId: 53,
+    isEditing: false,
   },
   {
     id: 3,
@@ -43,6 +54,7 @@ const passwordAccountData = [
     secretPictureName: 'wizard',
     secretPicturePath: '/wizard.jpg',
     sectionId: 53,
+    isEditing: false,
   }
 ];
 
@@ -59,6 +71,7 @@ const wordAccountData = [
     secretPictureName: 'wizard',
     secretPicturePath: '/wizard.jpg',
     sectionId: 53,
+    isEditing: false,
   },
   {
     id: 2,
@@ -72,6 +85,7 @@ const wordAccountData = [
     secretPictureName: 'wizard',
     secretPicturePath: '/wizard.jpg',
     sectionId: 53,
+    isEditing: false,
   },
   {
     id: 3,
@@ -101,6 +115,7 @@ const pictureAccountData = [
     secretPictureName: 'wizard',
     secretPicturePath: 'http://code.org/images/password_images/pirate_thumb@2x.png',
     sectionId: 53,
+    isEditing: false,
   },
   {
     id: 2,
@@ -114,6 +129,7 @@ const pictureAccountData = [
     secretPictureName: 'wizard',
     secretPicturePath: 'http://code.org/images/password_images/pirate_thumb@2x.png',
     sectionId: 53,
+    isEditing: false,
   },
   {
     id: 3,
@@ -127,6 +143,7 @@ const pictureAccountData = [
     secretPictureName: 'wizard',
     secretPicturePath: 'http://code.org/images/password_images/pirate_thumb@2x.png',
     sectionId: 53,
+    isEditing: false,
   }
 ];
 
@@ -189,6 +206,7 @@ const cleverData = [
 ];
 
 export default storybook => {
+  const store = createStore(combineReducers({manageStudents}), initialState);
   storybook
     .storiesOf('ManageStudentsTable', module)
     .addStoryTable([
@@ -196,55 +214,65 @@ export default storybook => {
         name: 'Table for email accounts',
         description: 'Ability to edit/delete all data including the password',
         story: () => (
-          <ManageStudentsTable
-            studentData={passwordAccountData}
-            id={53}
-            loginType={SectionLoginType.email}
-          />
+          <Provider store={store}>
+            <UnconnectedManageStudentsTable
+              studentData={passwordAccountData}
+              id={53}
+              loginType={SectionLoginType.email}
+            />
+          </Provider>
         )
       },
       {
         name: 'Table for word accounts',
         description: 'Ability to edit/delete all data and reset the secret word',
         story: () => (
-          <ManageStudentsTable
-            studentData={wordAccountData}
-            id={53}
-            loginType={SectionLoginType.word}
-          />
+          <Provider store={store}>
+            <UnconnectedManageStudentsTable
+              studentData={wordAccountData}
+              id={53}
+              loginType={SectionLoginType.word}
+            />
+          </Provider>
         )
       },
       {
         name: 'Table for picture accounts',
         description: 'Ability to edit/delete all data and reset the secret picture',
         story: () => (
-          <ManageStudentsTable
-            studentData={pictureAccountData}
-            id={53}
-            loginType={SectionLoginType.picture}
-          />
+          <Provider store={store}>
+            <UnconnectedManageStudentsTable
+              studentData={pictureAccountData}
+              id={53}
+              loginType={SectionLoginType.picture}
+            />
+          </Provider>
         )
       },
       {
         name: 'Table for Google accounts',
         description: 'Read only table',
         story: () => (
-          <ManageStudentsTable
-            studentData={googleData}
-            id={53}
-            loginType={SectionLoginType.google_classroom}
-          />
+          <Provider store={store}>
+            <UnconnectedManageStudentsTable
+              studentData={googleData}
+              id={53}
+              loginType={SectionLoginType.google_classroom}
+            />
+          </Provider>
         )
       },
       {
         name: 'Table for Clever accounts',
         description: 'Ready only table',
         story: () => (
-          <ManageStudentsTable
-            studentData={cleverData}
-            id={53}
-            loginType={SectionLoginType.clever}
-          />
+          <Provider store={store}>
+            <UnconnectedManageStudentsTable
+              studentData={cleverData}
+              id={53}
+              loginType={SectionLoginType.clever}
+            />
+          </Provider>
         )
       },
     ]);
