@@ -1,6 +1,7 @@
 require 'test_helper'
 
 class RegionalPartnersSchoolDistrictTest < ActiveSupport::TestCase
+  # rubocop:disable Style/NumericLiterals
   test "regional partners school districts initialized from tsv" do
     # populate school_districts and regional_partners, which we depend on as foreign keys
     RegionalPartner.find_or_create_all_from_tsv('test/fixtures/regional_partners.tsv')
@@ -14,7 +15,6 @@ class RegionalPartnersSchoolDistrictTest < ActiveSupport::TestCase
     assert_equal partner.name, 'A+ College Ready'
     assert_equal partner.group, 1
 
-    # rubocop:disable Style/NumericLiterals
     district = SchoolDistrict.find(4800014)
     partner = district.regional_partners.first
     assert_nil partner
@@ -25,4 +25,5 @@ class RegionalPartnersSchoolDistrictTest < ActiveSupport::TestCase
     assert_equal partner.name, 'Center for STEM Education, The University of Texas at Austin'
     assert_equal partner.group, 2
   end
+  # rubocop:enable Style/NumericLiterals
 end
