@@ -1,8 +1,13 @@
 module Teachercon1819RegistrationConstants
   TEACHER_SEAT_ACCEPTANCE_OPTIONS = {
-    accept: "Yes, I accept my seat in the Professional Learning Program",
-    decline: "No, I decline my seat in the Professional Learning Program.",
-    withdraw_date: "Yes, I want to participate, but I'm unable to attend my assigned summer workshop date. Please place me on your waitlist. I understand that I am not guaranteed a space in a different summer workshop.",
-    withdraw_other: "Yes, I want to participate, but I'm not able to for a different reason. Please place me on your waitlist. I understand that I am not guaranteed a space in a different summer workshop.",
-  }
+    accept: "Yes, I accept my seat in the Professional Learning Program".freeze,
+    decline: "No, I decline my seat in the Professional Learning Program.".freeze,
+    waitlist_date: "Yes, I want to participate, but I'm unable to attend my assigned summer workshop date. Please place me on your waitlist. I understand that I am not guaranteed a space in a different summer workshop.".freeze,
+    waitlist_other: "Yes, I want to participate, but I'm not able to for a different reason. Please place me on your waitlist. I understand that I am not guaranteed a space in a different summer workshop.".freeze,
+  }.freeze
+
+  TEXT_FIELDS = {
+    other_please_explain: 'Other (Please Explain):'.freeze,
+    food_allergy: 'Food Allergy'.freeze
+  }.freeze
 end
