@@ -57,17 +57,17 @@ def css_retina?(is_retina = true)
   css_query_parts.map {|q| "#{!is_retina ? 'not all and ' : ''}(#{q})"}.join(', ')
 end
 
-# Returns a concatenated, minified CSS string from all CSS files in the given paths.
+# Returns a concatenated, minified CSS string from all CSS files in the given paths,
+# along with a digest of same.
 def combine_css(*paths)
   files = paths.map {|path| Dir.glob(pegasus_dir('sites.v3', request.site, path, '*.css'))}.flatten
-  css_last_modified = Time.at(0)
   css = files.sort_by(&File.method(:basename)).map do |i|
-    css_last_modified = [css_last_modified, File.mtime(i)].max
     IO.read(i)
   end.join("\n\n")
   css_min = Sass::Engine.new(css,
     syntax: :scss,
     style: :compressed
   ).render
-  [css_min, css_last_modified]
+  digest = Digest::MD5.hexdigest(css_min)
+  [css_min, digest]
 end
